feat: Circle CI config
Removed travis config

Updated repo reference

feat: circle ci badge

feat: added shebang

fix

Remove map action test

+++ upload-ipa.sh
@@ -1,42 +1,23 @@
-#!/bin/sh
-if [[ "$TRAVIS_PULL_REQUEST" != "false" ]]; then
-    echo "This is a pull request. No deployment will be done."
-    exit 0
+#!/bin/bash
+
+if [[ $CIRCLE_BRANCH != pull* ]]
+then
+        git config --global user.name "chashmeetsingh"
+        git config --global user.email "chashmeetsingh@gmail.com"
+        git clone --quiet --branch=ipa https://chashmeetsingh:$GITHUB_API_KEY@github.com/fossasia/susi_iOS ipa > /dev/null
+        ls
+        rm -rf ipa/*.*
+        cp -r $HOME/Library/Developer/Xcode/DerivedData/Susi-*/Build/Products/Debug-iphonesimulator/Susi.app ipa/Susi.app
+        cd ipa
+        ls
+
+        git checkout --orphan workaround
+        git add Susi.app
+
+        git commit -am "[Circle CI] Update Susi.app"
+
+        git branch -D ipa
+        git branch -m ipa
+
+        git push origin ipa --force --quiet > /dev/null
 fi
-if [[ "$TRAVIS_BRANCH" != "master" ]]; then
-    echo "Testing on a branch other than master. No deployment will be done."
-    exit 0
-fi
-
-APPNAME="Susi"
-OUTPUTDIR="$PWD/build/Debug-iphoneos"
-
-git config --global user.email "noreply@travis.com"
-git config --global user.name "Travis CI"
-
-echo "********************"
-echo "*    Uploading     *"
-echo "********************"
-
-#create a new directory that will contain out generated apk
-mkdir $HOME/buildiPA/
-
-#copy .ipa from build folder and README.md to the folder just created
-cp -R @$OUTPUTDIR/$APPNAME.ipa $HOME/buildiPA/
-cp -R README.md $HOME/buildiPA/
-
-git clone --quiet --branch=apk https://fossasia:$GITHUB_API_KEY@github.com/fossasia/susi_iOS ipa > /dev/null
-
-cd ipa
-cp -Rf $HOME/buildiPA/*  ./
-
-git checkout --orphan workaround
-git add -A
-
-git commit -am "Travis build pushed [skip ci]"
-
-git branch -D ipa
-git branch -m ipa
-
-#push to the branch apk
-git push origin ipa --force --quiet> /dev/null
